Remove one nesting level from album playlist.
