+++ src/renderer/components/playlist/AlbumPlaylistItem.jsx
@@ -8,11 +8,8 @@ var DragSource = require('react-dnd').DragSource
 var DropTarget = require('react-dnd').DropTarget
 var cx = require('classnames')
 var key = require('keymaster')
-var moment = require('moment')
-require("moment-duration-format")
 
 var DragDropConstants = require('../../constants/DragDropConstants')
-var AlbumTracklistItem = require('./AlbumTracklistItem.jsx')
 var KeyboardFocusActions = require('../../actions/KeyboardFocusActions')
 var ContextMenuActions = require('../../actions/ContextMenuActions')
 var KeyboardNameSpaceConstants = require('../../constants/KeyboardNameSpaceConstants')
@@ -62,39 +59,10 @@ var AlbumPlaylistItem = React.createClass({
       cover: null
     }
   },
-  formatTime: function(time){
-    return moment.duration(time, "seconds").format("mm:ss", { trim: false })
-  },
   componentWillMount: function(){
     playa.coverLoader.load(this.props.album)
       .then(this.updateCover)
       .catch((err)=>{})
-  },
-  renderTracklist: function(){
-    var isMultiple = this.props.album.isMultiple()
-    var renderedTracklist = []
-    this.props.album.tracks.forEach((track, index)=>{
-      if(isMultiple && track.metadata.track == 1){
-        renderedTracklist.push((
-          <li key={track.id + '_disc_' + track.metadata.disk.no } className="disc-number">Disc {track.metadata.disk.no}</li>
-        ))
-      }
-      renderedTracklist.push(
-        <AlbumTracklistItem
-          key={track.id}
-          itemKey={track.id}
-          album={this.props.album}
-          track={track}
-          index={index}
-          selected={this.props.selection.indexOf(track.id) > -1}
-          isPlaying={track.id == this.props.currentTrack.id}
-          handleClick={this.handleTracklistClick}
-          handleDoubleClick={this.handleTracklistDoubleClick}/>
-      )
-    })
-    return (
-      <ol className="list-unstyled tracklist">{ renderedTracklist }</ol>
-    )
   },
   render: function() {
     var isPlaying = this.props.album.contains(this.props.currentTrack.id)
@@ -112,15 +80,12 @@ var AlbumPlaylistItem = React.createClass({
       'menuOpened'  : !!this.props.isMenuOpened
     })
     return this.props.connectDragSource(this.props.connectDropTarget(
-      <li className={classes} onClick={this.handleClick} onDoubleClick={this.handleDoubleClick} onContextMenu={this.handleMenuLinkClick} data-id={this.props.album.id} style={{opacity}}>
-        <header>
-          <div className={coverClasses} style={coverStyle}></div>
-          <span className="artist">{this.props.album.getArtist()}</span><br/>
-          <span className="title">{this.props.album.getTitle()} { (isPlaying && !this.props.isOpened) ? <i className="fa fa-fw fa-volume-up"></i> : null }</span>
-          <a href="#" className="menu-link sidebar-offset" onClick={this.handleMenuLinkClick}><i className="fa fa-fw fa-ellipsis-h"></i></a>
-          <span className="year sidebar-offset">{this.props.album.getYear()}</span>
-        </header>
-        { this.props.isOpened ? this.renderTracklist() : null }
+      <li className={classes} onClick={this.handleClick} onDoubleClick={this.handleDoubleClick} onContextMenu={this.handleMenuLinkClick} style={{opacity}} data-id={this.props.album.id}>
+        <div className={coverClasses} style={coverStyle}></div>
+        <span className="artist">{this.props.album.getArtist()}</span><br/>
+        <span className="title">{this.props.album.getTitle()} { (isPlaying && !this.props.isOpened) ? <i className="fa fa-fw fa-volume-up"></i> : null }</span>
+        <a href="#" className="menu-link sidebar-offset" onClick={this.handleMenuLinkClick}><i className="fa fa-fw fa-ellipsis-h"></i></a>
+        <span className="year sidebar-offset">{this.props.album.getYear()}</span>
       </li>
     ))
   },
@@ -133,25 +98,12 @@ var AlbumPlaylistItem = React.createClass({
       KeyboardNameSpaceConstants.ALBUM_PLAYLIST
     )
   },
-  handleTracklistClick: function(event, item){
-    event.stopPropagation()
-    this.props.handleClick(event, item)
-  },
-  handleTracklistDoubleClick: function(event, item){
-    event.stopPropagation()
-    this.props.playTrack(this.props.album, item.props.track.id)
-  },
   handleDoubleClick: function(event){
     event.stopPropagation()
     this.props.playTrack(this.props.album, this.props.album.tracks[0].id)
   },
   handleClick: function(event){
     this.props.handleClick(event, this)
-  },
-  handleEnterKeyPress: function(event){
-    if(this.state.selectedTrack > -1){
-      this.props.playTrack(this.props.album, this.props.album.tracks[this.state.selectedTrack].id)
-    }
   },
   updateCover: function(cover){
     this.setState({ cover: cover })
