+++ src/styles/_variables.styl
@@ -42,7 +42,7 @@ $tabsBgActiveColor = rgba(white, .075)
 // Playlist
 $playlistItemHeight = 2 * $baseFontSize
 $playlistFontSize = .9 * $baseFontSize
-$albumItemClosedHeight = 4 * $baseFontSize
+$albumItemHeight = 4 * $baseFontSize
 $albumHoverBackgroundColor = rgba(white, .05)
 $albumSelectedBackgroundColor = rgba(white, .075)
 $albumPlayingBackgroundColor = $mainColor
@@ -57,7 +57,7 @@ $albumTracklistLineHeight = 2 * $baseFontSize
 $albumItemMenuLinkBgColor = rgba(white, .1)
 
 // Playlist Drop Area
-$dropAreaHeight = 3 * $albumItemClosedHeight
+$dropAreaHeight = 3 * $albumItemHeight
 $dropAreaHoverBgColor = $albumHoverBackgroundColor
 $dropAreaTextColor = rgba(white, .25)
 $dropAreaFontSize = 2 * $baseFontSize
